+++ ui/features/gradebook/react/default_gradebook/__tests__/GradebookSort4.test.js
@@ -17,110 +17,7 @@
  */
 
 import {createGradebook, setFixtureHtml} from './GradebookSpecHelper'
-import GradebookApi from '../apis/GradebookApi'
 import sinon from 'sinon'
-
-describe('Gradebook#sortGridRows', () => {
-  let gradebook
-  let sandbox
-
-  beforeEach(() => {
-    sandbox = sinon.createSandbox()
-    window.ENV = {
-      current_user_id: '1',
-      context_id: '1',
-      GRADEBOOK_OPTIONS: {
-        custom_columns: [],
-        grading_schemes: [],
-        settings_update_url: '/courses/1/gradebook_settings',
-      },
-      FEATURES: {instui_nav: false},
-    }
-    gradebook = createGradebook()
-    gradebook.gridData.rows = [
-      {id: '3', sortable_name: 'Z', total_grade: {score: 15}, assignment_2301: {score: 10}},
-      {id: '4', sortable_name: 'A', total_grade: {score: 10}, assignment_2301: {score: 15}},
-    ]
-    gradebook.setAssignments({
-      2301: {
-        id: '2301',
-        grading_type: 'points',
-        name: 'Assignment 1',
-        published: true,
-        submission_types: ['online_text_entry'],
-      },
-    })
-    gradebook.gridDisplaySettings.viewUngradedAsZero = false
-    sandbox.stub(gradebook.gradebookGrid, 'updateColumns')
-    sandbox.stub(gradebook.gradebookGrid.gridSupport.columns, 'updateColumnHeaders')
-  })
-
-  afterEach(() => {
-    sandbox.restore()
-    window.ENV = undefined
-  })
-
-  test('uses the saved sort setting for student column sorting', () => {
-    gradebook.setSortRowsBySetting('student_name', 'ascending')
-    gradebook.sortGridRows()
-    const [firstRow, secondRow] = gradebook.gridData.rows
-    expect(firstRow.id).toBe('4')
-    expect(secondRow.id).toBe('3')
-  })
-
-  test.skip('optionally sorts by a custom column', () => {
-    gradebook.setSortRowsBySetting('custom_col_2301', 'ascending')
-    gradebook.sortGridRows()
-    const [firstRow, secondRow] = gradebook.gridData.rows
-    expect(firstRow.id).toBe('4')
-    expect(secondRow.id).toBe('3')
-  })
-
-  test('uses the saved sort setting for custom column sorting', () => {
-    gradebook.setSortRowsBySetting('custom_col_2301', 'descending')
-    gradebook.sortGridRows()
-    const [firstRow, secondRow] = gradebook.gridData.rows
-    expect(firstRow.id).toBe('3')
-    expect(secondRow.id).toBe('4')
-  })
-
-  test('optionally sorts by an assignment column', () => {
-    gradebook.setSortRowsBySetting('assignment_2301', 'ascending')
-    gradebook.sortGridRows()
-    const [firstRow, secondRow] = gradebook.gridData.rows
-    expect(firstRow.id).toBe('3')
-    expect(secondRow.id).toBe('4')
-  })
-
-  test.skip('uses the saved sort setting for assignment sorting', () => {
-    gradebook.setSortRowsBySetting('assignment_2301', 'descending')
-    gradebook.sortGridRows()
-    const [firstRow, secondRow] = gradebook.gridData.rows
-    expect(firstRow.id).toBe('4')
-    expect(secondRow.id).toBe('3')
-  })
-
-  test.skip('optionally sorts by the total grade column', () => {
-    gradebook.setSortRowsBySetting('total_grade', 'ascending')
-    gradebook.sortGridRows()
-    const [firstRow, secondRow] = gradebook.gridData.rows
-    expect(firstRow.id).toBe('4')
-    expect(secondRow.id).toBe('3')
-  })
-
-  test('uses the saved sort setting for total grade sorting', () => {
-    gradebook.setSortRowsBySetting('total_grade', 'descending')
-    gradebook.sortGridRows()
-    const [firstRow, secondRow] = gradebook.gridData.rows
-    expect(firstRow.id).toBe('3')
-    expect(secondRow.id).toBe('4')
-  })
-
-  test('updates the column headers after sorting', () => {
-    gradebook.sortGridRows()
-    expect(gradebook.gradebookGrid.gridSupport.columns.updateColumnHeaders.callCount).toBe(1)
-  })
-})
 
 describe('Gradebook#getColumnSortSettingsViewOptionsMenuProps', () => {
   let gradebook
@@ -264,87 +161,5 @@ describe('Gradebook#getColumnSortSettingsViewOptionsMenuProps', () => {
     expect(gradebook.arrangeColumnsBy.firstCall.args).toEqual(
       expectedArgs('module_position', 'descending'),
     )
-  })
-})
-
-describe('when enhanced_gradebook_filters is enabled', () => {
-  let gradebook
-  let errorFn
-  let successFn
-  let saveUserSettingsStub
-  let sandbox
-  let $fixtures
-  let oldEnv
-
-  beforeEach(() => {
-    $fixtures = document.createElement('div')
-    document.body.appendChild($fixtures)
-    setFixtureHtml($fixtures)
-    sandbox = sinon.createSandbox()
-    oldEnv = window.ENV
-    window.ENV = {
-      FEATURES: {instui_nav: true},
-      current_user_id: '1',
-      context_id: '1',
-      GRADEBOOK_OPTIONS: {
-        custom_columns: [],
-        grading_schemes: [],
-      },
-    }
-    gradebook = createGradebook({
-      enhanced_gradebook_filters: true,
-    })
-    const assignment = {
-      id: '2301',
-      grading_type: 'points',
-      name: 'Assignment 1',
-      published: true,
-      submission_types: ['online_text_entry'],
-    }
-    gradebook.setAssignments({2301: assignment})
-    gradebook.setAssignmentsLoaded()
-
-    errorFn = sandbox.stub()
-    successFn = sandbox.stub()
-
-    saveUserSettingsStub = sandbox.stub(GradebookApi, 'saveUserSettings')
-  })
-
-  afterEach(() => {
-    if (gradebook) {
-      gradebook.destroy && gradebook.destroy()
-    }
-    $fixtures.remove()
-    saveUserSettingsStub.restore()
-    sandbox.restore()
-    window.ENV = oldEnv
-  })
-
-  test('calls the provided successFn if the request succeeds', async () => {
-    saveUserSettingsStub.resolves({})
-    await gradebook.saveSettings({}).then(successFn).catch(errorFn)
-    expect(successFn.callCount).toBe(1)
-    expect(errorFn.notCalled).toBeTruthy()
-  })
-
-  test('calls the provided errorFn if the request fails', async () => {
-    saveUserSettingsStub.rejects(new Error(':('))
-    await gradebook.saveSettings({}).then(successFn).catch(errorFn)
-    expect(errorFn.callCount).toBe(1)
-    expect(successFn.notCalled).toBeTruthy()
-  })
-
-  test('just returns if the request succeeds and no successFn is provided', async () => {
-    // QUnit.expect(0) is not needed in Jest
-    saveUserSettingsStub.resolves({})
-    await gradebook.saveSettings({})
-    // No assertions needed
-  })
-
-  test('throws an error if the request fails and no errorFn is provided', async () => {
-    // QUnit.expect(1) is not needed in Jest
-    saveUserSettingsStub.rejects(new Error('>:('))
-
-    await expect(gradebook.saveSettings({})).rejects.toThrow('>:(')
   })
 })
